Made subBanner sticky & alligned its width to main
