+++ html/landingpages/information.html
@@ -97,8 +97,8 @@
 		</aside>
 		
 		<main class="main-container">
-			<div class="main-container__subscribe-form">
-				<div class="main-container__subscribe-banner"><a href="./html/landingpages/information.html">Subscribe for architecture, art, and more</a>
+			<div class="main-container__subscribe">
+				<div class="main-container__subscribe-banner"><a href="/html/landingpages/information.html" class="main-container__subscribe-text">Subscribe<span> for architecture, art, and more</span></a>
 					<button class="main-container__subscribe-banner-close"></button>
 				</div>
 			</div>
