+++ js/main.js
@@ -1,9 +1,7 @@
 import Hover from './modules/Hover.js'
 import subscribeBanner from './modules/SubscribeBanner.js'
-import subscribeButton from './modules/SubscribeButton.js'
 
 Hover();
+subscribeBanner();
 Rederect();
 controlSubMenuMobile();
-subscribeButton();
-subscribeBanner();
